requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
jiter==0.7.1
numpy==2.1.3
openai==1.55.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==2.10.0
pydantic_core==2.27.0
pygame==2.6.1
pytesseract==0.3.13
python-telegram-bot==21.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
